Various bug fixes

Closes #978 and #976

See merge request openlp/openlp!384

+++ openlp/core/ui/themelayoutform.py
@@ -45,5 +45,5 @@ class ThemeLayoutForm(QtWidgets.QDialog, Ui_ThemeLayoutDialog):
         pixmap.setDevicePixelRatio(self.theme_display_label.devicePixelRatio())
         self.theme_display_label.setPixmap(pixmap)
         display_aspect_ratio = float(image.width()) / image.height()
-        self.theme_display_label.setFixedSize(400, 400 / display_aspect_ratio)
+        self.theme_display_label.setFixedSize(400, int(400 / display_aspect_ratio))
         return QtWidgets.QDialog.exec(self)
